+++ src/app/_components/Header.tsx
@@ -73,6 +73,9 @@ export default function RenderHeader() {
                   <li className="my-2 border-b border-gray-400 uppercase">
                     <a href="/learnflows">Learn Flows</a>
                   </li>
+                  <li className="my-2 border-b border-gray-400 uppercase">
+                    <a href="/viewflows">View Flows</a>
+                  </li>
                   {inDevelopment || (
                     <li className="my-2 border-b border-gray-400 uppercase">
                       <a href="/combos">Learn Combos</a>
